[Rename] management plugin
