--- src/plugins/management/kibana.json
@@ -1,8 +1,0 @@
-{
-  "id": "management",
-  "version": "kibana",
-  "server": true,
-  "ui": true,
-  "optionalPlugins": ["home"],
-  "requiredBundles": ["kibanaReact", "kibanaUtils", "home"]
-}
