+++ src/plugins/management/opensearch_dashboards.json
@@ -1,0 +1,8 @@
+{
+  "id": "management",
+  "version": "opensearchDashboards",
+  "server": true,
+  "ui": true,
+  "optionalPlugins": ["home"],
+  "requiredBundles": ["opensearchDashboardsReact", "opensearchDashboardsUtils", "home"]
+}
