+++ src/plugins/management/public/components/management_app/management_router.tsx
@@ -20,7 +20,7 @@
 import React, { memo } from 'react';
 import { Route, Router, Switch } from 'react-router-dom';
 import { EuiPageBody } from '@elastic/eui';
-import { AppMountParameters, ChromeBreadcrumb, ScopedHistory } from 'kibana/public';
+import { AppMountParameters, ChromeBreadcrumb, ScopedHistory } from 'opensearch-dashboards/public';
 import { ManagementAppWrapper } from '../management_app_wrapper';
 import { ManagementLandingPage } from '../landing';
 import { ManagementAppDependencies } from './management_app';
@@ -60,7 +60,7 @@ export const ManagementRouter = memo(
             path={'/'}
             component={() => (
               <ManagementLandingPage
-                version={dependencies.kibanaVersion}
+                version={dependencies.opensearchDashboardsVersion}
                 setBreadcrumbs={setBreadcrumbs}
               />
             )}
